Refactor resource cache: Update references to HashList and introduce resourcecachehashlist package, enhance RorClientInterface with GetOwnerref method
