+++ pkg/helpers/resourcecache/resourcecachehashlist/hashlist.go
@@ -1,4 +1,4 @@
-package hashlist
+package resourcecachehashlist
 
 import (
 	"fmt"
